Clear the job and the panel references from memory when both the query file and the viewer are closed
